+++ config/routes.rb
@@ -10,8 +10,8 @@ Rails.application.routes.draw do
   get 'doc/hpi', to: 'doc#hpi', as: 'hpi_doc'
   get 'doc/ukhpi', to: redirect('/app/ukhpi/doc', status: 302)
   get 'app/doc/ppd', to: 'doc#ppd', as: 'ppd_doc'
-  get 'doc/ppd', to: redirect('/app/doc/ppd', status: 301)
-  get 'app/root/doc/ppd', to: redirect('app/doc/ppd', status: 301)
+  get 'doc/ppd', to: redirect('/app/doc/ppd', status: 302)
+  get 'app/root/doc/ppd', to: redirect('app/doc/ppd', status: 302)
   get 'doc/ukhpi-dsd', to: redirect('/app/ukhpi/doc/ukhpi-dsd', status: 302)
   get 'doc/ukhpi-user-guide', to: redirect('/app/ukhpi/doc/ukhpi-user-guide', status: 302)
   get 'doc/accessibility', to: 'doc#accessibility'